test: fix module cli tests
