+++ x/bank/client/cli/tx_test.go
@@ -11,6 +11,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 	"github.com/cosmos/cosmos-sdk/testutil"
+	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/bank/client/cli"
 )
@@ -29,12 +30,12 @@ func (s *CLITestSuite) TestSendTxCmd() {
 	}
 
 	testCases := []struct {
-		name      string
-		ctxGen    func() client.Context
-		from, to  sdk.AccAddress
-		amount    sdk.Coins
-		extraArgs []string
-		expectErr bool
+		name         string
+		ctxGen       func() client.Context
+		from, to     sdk.AccAddress
+		amount       sdk.Coins
+		extraArgs    []string
+		expectErrMsg string
 	}{
 		{
 			"valid transaction",
@@ -48,7 +49,7 @@ func (s *CLITestSuite) TestSendTxCmd() {
 				sdk.NewCoin("photon", sdkmath.NewInt(40)),
 			),
 			extraArgs,
-			false,
+			"",
 		},
 		{
 			"invalid to Address",
@@ -62,7 +63,7 @@ func (s *CLITestSuite) TestSendTxCmd() {
 				sdk.NewCoin("photon", sdkmath.NewInt(40)),
 			),
 			extraArgs,
-			true,
+			"empty address string is not allowed",
 		},
 		{
 			"invalid coins",
@@ -73,25 +74,28 @@ func (s *CLITestSuite) TestSendTxCmd() {
 			accounts[0].Address,
 			nil,
 			extraArgs,
-			true,
+			"invalid coins",
 		},
 	}
 
 	for _, tc := range testCases {
 		tc := tc
 		s.Run(tc.name, func() {
+			args := append([]string{tc.from.String(), tc.to.String(), tc.amount.String()}, tc.extraArgs...)
+
 			ctx := svrcmd.CreateExecuteContext(context.Background())
-
 			cmd.SetContext(ctx)
-			cmd.SetArgs(append([]string{tc.from.String(), tc.to.String(), tc.amount.String()}, tc.extraArgs...))
-
+			cmd.SetArgs(args)
 			s.Require().NoError(client.SetCmdClientContextHandler(tc.ctxGen(), cmd))
 
-			err := cmd.Execute()
-			if tc.expectErr {
+			out, err := clitestutil.ExecTestCLICmd(tc.ctxGen(), cmd, args)
+			if tc.expectErrMsg != "" {
 				s.Require().Error(err)
+				s.Require().Contains(out.String(), tc.expectErrMsg)
 			} else {
 				s.Require().NoError(err)
+				msg := &sdk.TxResponse{}
+				s.Require().NoError(tc.ctxGen().Codec.UnmarshalJSON(out.Bytes(), msg), out.String())
 			}
 		})
 	}
@@ -112,13 +116,13 @@ func (s *CLITestSuite) TestMultiSendTxCmd() {
 	}
 
 	testCases := []struct {
-		name      string
-		ctxGen    func() client.Context
-		from      string
-		to        []string
-		amount    sdk.Coins
-		extraArgs []string
-		expectErr bool
+		name         string
+		ctxGen       func() client.Context
+		from         string
+		to           []string
+		amount       sdk.Coins
+		extraArgs    []string
+		expectErrMsg string
 	}{
 		{
 			"valid transaction",
@@ -135,7 +139,7 @@ func (s *CLITestSuite) TestMultiSendTxCmd() {
 				sdk.NewCoin("photon", sdkmath.NewInt(40)),
 			),
 			extraArgs,
-			false,
+			"",
 		},
 		{
 			"invalid from Address",
@@ -152,7 +156,7 @@ func (s *CLITestSuite) TestMultiSendTxCmd() {
 				sdk.NewCoin("photon", sdkmath.NewInt(40)),
 			),
 			extraArgs,
-			true,
+			"key not found",
 		},
 		{
 			"invalid recipients",
@@ -169,7 +173,7 @@ func (s *CLITestSuite) TestMultiSendTxCmd() {
 				sdk.NewCoin("photon", sdkmath.NewInt(40)),
 			),
 			extraArgs,
-			true,
+			"invalid bech32 string",
 		},
 		{
 			"invalid amount",
@@ -183,7 +187,7 @@ func (s *CLITestSuite) TestMultiSendTxCmd() {
 			},
 			nil,
 			extraArgs,
-			true,
+			"must send positive amount",
 		},
 	}
 
@@ -203,11 +207,14 @@ func (s *CLITestSuite) TestMultiSendTxCmd() {
 
 			s.Require().NoError(client.SetCmdClientContextHandler(tc.ctxGen(), cmd))
 
-			err := cmd.Execute()
-			if tc.expectErr {
+			out, err := clitestutil.ExecTestCLICmd(tc.ctxGen(), cmd, args)
+			if tc.expectErrMsg != "" {
 				s.Require().Error(err)
+				s.Require().Contains(out.String(), tc.expectErrMsg)
 			} else {
 				s.Require().NoError(err)
+				msg := &sdk.TxResponse{}
+				s.Require().NoError(tc.ctxGen().Codec.UnmarshalJSON(out.Bytes(), msg), out.String())
 			}
 		})
 	}
