+++ x/evidence/client/cli/query_test.go
@@ -1,7 +1,6 @@
 package cli_test
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"io"
@@ -10,7 +9,6 @@ import (
 
 	"cosmossdk.io/x/evidence"
 	"cosmossdk.io/x/evidence/client/cli"
-	abci "github.com/cometbft/cometbft/abci/types"
 	rpcclientmock "github.com/cometbft/cometbft/rpc/client/mock"
 	"github.com/stretchr/testify/require"
 
@@ -19,7 +17,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/crypto/keyring"
 	svrcmd "github.com/cosmos/cosmos-sdk/server/cmd"
 	clitestutil "github.com/cosmos/cosmos-sdk/testutil/cli"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	testutilmod "github.com/cosmos/cosmos-sdk/types/module/testutil"
 )
 
@@ -37,51 +34,25 @@ func TestGetQueryCmd(t *testing.T) {
 
 	testCases := map[string]struct {
 		args           []string
-		ctxGen         func() client.Context
-		expCmdOutput   string
 		expectedOutput string
-		expectErr      bool
+		expectErrMsg   string
 	}{
-		"non-existent evidence": {
-			[]string{"DF0C23E8634E480F84B9D5674A7CDC9816466DEC28A3358F73260F68D28D7660"},
-			func() client.Context {
-				bz, _ := encCfg.Codec.Marshal(&sdk.TxResponse{})
-				c := clitestutil.NewMockCometRPC(abci.ResponseQuery{
-					Value: bz,
-				})
-				return baseCtx.WithClient(c)
-			},
-			"DF0C23E8634E480F84B9D5674A7CDC9816466DEC28A3358F73260F68D28D7660",
+		"invalid args": {
+			[]string{"foo", "bar"},
 			"",
-			true,
+			"accepts at most 1 arg(s)",
 		},
 		"all evidence (default pagination)": {
 			[]string{},
-			func() client.Context {
-				bz, _ := encCfg.Codec.Marshal(&sdk.TxResponse{})
-				c := clitestutil.NewMockCometRPC(abci.ResponseQuery{
-					Value: bz,
-				})
-				return baseCtx.WithClient(c)
-			},
-			"",
 			"evidence: []\npagination: null",
-			false,
+			"",
 		},
 		"all evidence (json output)": {
 			[]string{
 				fmt.Sprintf("--%s=json", flags.FlagOutput),
 			},
-			func() client.Context {
-				bz, _ := encCfg.Codec.Marshal(&sdk.TxResponse{})
-				c := clitestutil.NewMockCometRPC(abci.ResponseQuery{
-					Value: bz,
-				})
-				return baseCtx.WithClient(c)
-			},
-			"",
 			`{"evidence":[],"pagination":null}`,
-			false,
+			"",
 		},
 	}
 
@@ -89,31 +60,21 @@ func TestGetQueryCmd(t *testing.T) {
 		tc := tc
 
 		t.Run(name, func(t *testing.T) {
-			var outBuf bytes.Buffer
-
-			clientCtx := tc.ctxGen().WithOutput(&outBuf)
 			ctx := svrcmd.CreateExecuteContext(context.Background())
 
 			cmd := cli.GetQueryCmd()
-			cmd.SetOut(io.Discard)
-			require.NotNil(t, cmd)
-
 			cmd.SetContext(ctx)
 			cmd.SetArgs(tc.args)
-
-			require.NoError(t, client.SetCmdClientContextHandler(clientCtx, cmd))
-
-			if len(tc.args) != 0 {
-				require.Contains(t, fmt.Sprint(cmd), tc.expCmdOutput)
-			}
+			require.NoError(t, client.SetCmdClientContextHandler(baseCtx, cmd))
 
 			out, err := clitestutil.ExecTestCLICmd(baseCtx, cmd, tc.args)
-			if tc.expectErr {
+			if tc.expectErrMsg != "" {
 				require.Error(t, err)
+				require.Contains(t, err.Error(), tc.expectErrMsg)
 			} else {
 				require.NoError(t, err)
 			}
-			require.Contains(t, fmt.Sprint(cmd), "evidence [] [] Query for evidence by hash or for all (paginated) submitted evidence")
+
 			require.Contains(t, strings.TrimSpace(out.String()), tc.expectedOutput)
 		})
 	}
